[fix]사용자 마이페이지 - order 상세보기 페이지 fold기능 추가
